Restructered GefDilemma --> finished the work --> please check for faults
Thought we had to implement small functions for every functionality but
found out that we do not need to do that, so everything takes place in
the "spiele" function.
Removed the unnecessary parts of the code. Test cases are running and
returning correct output. Tested all different  strategies --> works
feedback, improvement suggestions?